Refactored student page
