+++ app/src/pages/student/option-parameters/index.ts
@@ -1,0 +1,1 @@
+export { OptionParameters } from './option-parameters';
