+++ app/src/pages/student/option-parameters/option-parameters.module.scss
@@ -1,0 +1,9 @@
+.sectionValue {
+  display: grid;
+  grid-template-columns: 3fr 3fr 1fr;
+  gap: 2rem;
+}
+
+.button {
+  background: var(--dark-color);
+}
