+++ app/src/pages/student/option/constants.ts
@@ -1,0 +1,16 @@
+export const OPTION = {
+  user: 'Иванов Иван Иванович',
+  optionCode: 'Jhbsr234hrbfAJDXNCBkjwh43bJSDB==',
+  tasks: [
+    {
+      id: 5,
+      amount: 1
+    },
+    {
+      id: 2,
+      amount: 1
+    }
+  ],
+  userHash: '309ecc489c12d6eb4cc40f50c902f2b4d0ed77ee511a7c7a9bcd3ca86' +
+    'd4cd86f4e0b87b20eabe88305d65fd0ed5d93b81de5b5e76d03dcc48e5286e044b7f3'
+};
