+++ app/src/pages/student/option/index.ts
@@ -1,0 +1,1 @@
+export { Option } from './option';
