+++ app/src/pages/student/option/option.tsx
@@ -1,0 +1,3 @@
+import { type FC } from 'react';
+
+export const Option: FC = () => <div />;
